// src/generated/static/types/admin/permission.ts
export type AdminPermission =
  | 'admin.create'
  | 'admin.delete'
  | 'admin.list'
  | 'admin.log.list'
  | 'admin.role.create'
  | 'admin.role.list'
  | 'admin.toggle'
  | 'admin.update'
  | 'home.dashboard.view';

export type AdminPermissionGroup =
  | 'admin'
  | 'admin.log'
  | 'admin.role'
  | 'home'
  | 'home.dashboard';